feat: Almost implement AddPostPage function
I implemented subject selection button, textFields, rule, image button
and complete posting button. I use TextEditingController, so Text in the
title and content textField can be printed on the console android
studio.

When subject is not selected or title or content is not entered, the
posting completion button should be implemented without being pressed.

When image button is pressed, the funtion should alse be created.